Modified the the section "Partnering with the career development office" to address questions from Reviewers 1 and 3 about team formation.
I did not add a full section on "Team Formation" because I do not think we
contributed anything new in this domain. I looked at the papers that Reviewer
1 suggested referencing with regards to team formation and I did not find them
relevant.

I moved the footnote in response to Reviewer 5's comments, but I wasn't there
when the inventories were dropped from the curriculum so I can't comment further
on our reasons for doing so or what took their place in fostering discussion of
diversity within teams.